Biscuit – Compiler Warning Cleanup (#16)
# Biscuit – Compiler Warning Cleanup

This patch silences every warning emitted from Biscuit's own source files
during the build.

* **`-Wunused-parameter` (PG_FUNCTION_ARGS, `fcinfo`)**
  `biscuit.c`: `biscuit_has_roaring`, `biscuit_version`,
  `biscuit_build_info_json`, `biscuit_handler`,
  `biscuit_like_support`, and `biscuit_roaring_version`
  don't use `fcinfo`. Added explicit `(void) fcinfo;` to each.

* **`-Wunused-parameter` (callback signatures)**
  `biscuit_cache.c`: `biscuit_relcache_callback` doesn't use `arg`;
  `biscuit_module_unload_callback` doesn't use `code`/`datum`. Added
  `(void)` casts — both are PostgreSQL callback signatures we can't
  change.

* **`-Wunused-parameter` (reserved-for-future-use parameters)**
  `biscuit_utf8.c`: `biscuit_datum_to_text`'s `outfunc` parameter is
  currently unused (reserved for future type support). Added `(void)`
  cast + comment.
  `biscuit_scan.c`: `biscuit_rescan_multicolumn` and
  `biscuit_rescan_multicolumn_fallback` don't use `orderbys`/`norderbys`
  (Biscuit has no ORDER BY support yet); `biscuit_gettuple` doesn't use
  `dir` (results are always returned in build order). Added `(void)`
  casts.
  `biscuit_pattern.c`: `biscuit_build_query_plan`'s `idx` parameter is
  unused (reserved for future cardinality-aware planning). Added `(void)`
  cast.

* **`-Wunused-variable`**
  `biscuit_scan.c`: moved the declaration of the local variable (`j`) to
  the scope where it is actually used, eliminating the
  `-Wunused-variable` warning while preserving the existing logic.


* **`-Wsign-compare` (tombstone bitmap fallback path)**
  In the non-`HAVE_ROARING` fallback bitset implementation, the inline
  `bl < idx->tombstones->num_blocks` check compares a `uint32_t` (`bl`)
  against an `int` (`num_blocks`). Cast `bl` to `int` at each of the 7
  call sites so the comparison is signed/signed:

  * `biscuit.c` (`biscuit_index_stats`)
  * `biscuit_index.c` (`biscuit_bulkdelete`)
  * `biscuit_preload.c` (`biscuit_fallback_scan`)
  * `biscuit_pattern.c` (`biscuit_query_pattern`, x2;
    `biscuit_query_pattern_ilike`, x2)

These changes eliminate the remaining compiler warnings originating from
Biscuit's source files without altering runtime behavior.

---

+++ src/biscuit_index.c
@@ -1478,7 +1478,7 @@ biscuit_bulkdelete(IndexVacuumInfo *info,
 #ifdef HAVE_ROARING
         already_tombstoned = roaring_bitmap_contains(idx->tombstones, (uint32_t) i);
 #else
-        { uint32_t bl = i >> 6, bt = i & 63; already_tombstoned = (bl < idx->tombstones->num_blocks && (idx->tombstones->blocks[bl] & (1ULL << bt))); }
+        { uint32_t bl = i >> 6, bt = i & 63; already_tombstoned = ((int) bl < idx->tombstones->num_blocks && (idx->tombstones->blocks[bl] & (1ULL << bt))); }
 #endif
         if (already_tombstoned) continue;
